dynamic pk manager
