+++ packages/create-video/package.json
@@ -22,7 +22,6 @@
 		"degit": "2.8.4",
 		"execa": "5.1.1",
 		"fs-extra": "9.1.0",
-		"prompts": "2.4.1",
 		"strip-ansi": "6.0.1"
 	},
 	"devDependencies": {
@@ -30,7 +29,6 @@
 		"@types/degit": "^2.8.3",
 		"@types/fs-extra": "^9.0.11",
 		"@types/node": "^16.7.5",
-		"@types/prompts": "^2.0.12",
 		"eslint": "^7.25.0",
 		"prettier-plugin-organize-imports": "^1.1.1",
 		"ts-node": "^9.1.1",
